feat: simplify event caller
